Add new instance handler

+++ Users.cs
@@ -56,8 +56,9 @@ namespace Proyek_UAS
         private void Back_Button_Click(object sender, EventArgs e)
         {
             this.Hide();
-            Form Home = new Home();
-            Home.Show();
+            //Form Home = new Home();
+            //Home.Visible = true;
+            FormHandler.Home.Show();
         }
 
         //Add Users
@@ -148,19 +149,27 @@ namespace Proyek_UAS
         //Delete Users
         private void deleteButton_Click(object sender, EventArgs e)
         {
-            var confirmResult = MessageBox.Show("Are you sure you want to delete this user?", "Confirmation", MessageBoxButtons.YesNo);
-            if (confirmResult == DialogResult.Yes)
+            var username = dataGridView1.SelectedCells[2].Value.ToString();
+            if (FormHandler.Username == username)
             {
-                var username = dataGridView1.SelectedCells[2].Value.ToString();
-                SqlCommand cmd = con.CreateCommand();
-                cmd.CommandType = CommandType.Text;
-                cmd.CommandText = "DELETE FROM Users WHERE username='" + username + "';";
-                cmd.ExecuteNonQuery();
+                MessageBox.Show("You can't delete yourself!");
+            } else
+            {
+                var confirmResult = MessageBox.Show("Are you sure you want to delete this user?", "Confirmation", MessageBoxButtons.YesNo);
+                if (confirmResult == DialogResult.Yes)
+                {
+                    SqlCommand cmd = con.CreateCommand();
+                    cmd.CommandType = CommandType.Text;
+                    cmd.CommandText = "DELETE FROM Users WHERE username='" + username + "';";
+                    cmd.ExecuteNonQuery();
 
-                display();
+                    display();
 
-                MessageBox.Show("User deleted!");
+                    MessageBox.Show("User deleted!");
+                }
             }
+                ;
+
         }
     }
 }
